moved repository interfaces to DataAbstractions folder
