+++ src/BoltOn.Data.MartenDb/QueryRepository.cs
@@ -5,6 +5,7 @@ using System.Linq.Expressions;
 using System.Threading.Tasks;
 using Marten;
 using System.Linq;
+using BoltOn.DataAbstractions.MartenDb;
 
 namespace BoltOn.Data.MartenDb
 {
