+++ src/BoltOn/DataAbstractions/EF/IQueryRepository.cs
@@ -4,7 +4,7 @@ using System.Linq.Expressions;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace BoltOn.Data.EF
+namespace BoltOn.DataAbstractions.EF
 {
     public interface IQueryRepository<TEntity> where TEntity : class
     {
