+++ src/BoltOn/DataAbstractions/EF/IRepository.cs
@@ -1,16 +1,15 @@
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace BoltOn.Data.MartenDb
+namespace BoltOn.DataAbstractions.EF
 {
-	public interface IRepository<TEntity> : IQueryRepository<TEntity>
+    public interface IRepository<TEntity> : IQueryRepository<TEntity>
        where TEntity : class
     {
         Task<TEntity> AddAsync(TEntity entity, CancellationToken cancellationToken = default);
         Task<IEnumerable<TEntity>> AddAsync(IEnumerable<TEntity> entities, CancellationToken cancellationToken = default);
         Task UpdateAsync(TEntity entity, CancellationToken cancellationToken = default);
         Task DeleteAsync(object id, CancellationToken cancellationToken = default);
-        Task<object[]> StoreAsync(CancellationToken cancellationToken = default, params object[] objects);
     }
 }
