+++ tests/BoltOn.Tests/Other/StudentRepository.cs
@@ -1,5 +1,6 @@
 using BoltOn.Data;
 using BoltOn.Data.EF;
+using BoltOn.DataAbstractions.EF;
 
 namespace BoltOn.Tests.Other
 {
